+++ core/api/typescript/src/lib/enclaves/grpc_node_api_container_client.ts
@@ -9,6 +9,7 @@ import {
     RepartitionArgs,
     WaitForHttpGetEndpointAvailabilityArgs,
     WaitForHttpPostEndpointAvailabilityArgs,
+    StartServiceResponse,
     GetServicesResponse,
     LoadModuleArgs,
     UnloadModuleArgs,
@@ -24,12 +25,7 @@ import {
     StoreWebFilesArtifactArgs,
     StoreWebFilesArtifactResponse,
     StoreFilesArtifactFromServiceArgs,
-    StoreFilesArtifactFromServiceResponse,
-    GetServicesArgs,
-    GetModulesArgs,
-    UnloadModuleResponse,
-    ServiceInfo,
-    ModuleInfo,
+    StoreFilesArtifactFromServiceResponse, GetServicesArgs, GetModulesArgs, UnloadModuleResponse,
 } from "../../kurtosis_core_rpc_api_bindings/api_container_service_pb";
 import type { ApiContainerServiceClient as ApiContainerServiceClientNode } from "../../kurtosis_core_rpc_api_bindings/api_container_service_grpc_pb";
 import { GenericApiContainerClient } from "./generic_api_container_client";
@@ -50,9 +46,9 @@ export class GrpcNodeApiContainerClient implements GenericApiContainerClient {
         return this.enclaveId;
     }
 
-    public async loadModule(loadModuleArgs: LoadModuleArgs): Promise<Result<ModuleInfo, Error>> {
-        const loadModulePromise: Promise<Result<ModuleInfo, Error>> = new Promise((resolve, _unusedReject) => {
-            this.client.loadModule(loadModuleArgs, (error: ServiceError | null, response?: ModuleInfo) => {
+    public async loadModule(loadModuleArgs: LoadModuleArgs): Promise<Result<null, Error>> {
+        const loadModulePromise: Promise<Result<google_protobuf_empty_pb.Empty, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.loadModule(loadModuleArgs, (error: ServiceError | null, response?: google_protobuf_empty_pb.Empty) => {
                 if (error === null) {
                     if (!response) {
                         resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
@@ -64,13 +60,12 @@ export class GrpcNodeApiContainerClient implements GenericApiContainerClient {
                 }
             })
         });
-        const loadModulePromiseResult: Result<ModuleInfo, Error> = await loadModulePromise;
+        const loadModulePromiseResult: Result<google_protobuf_empty_pb.Empty, Error> = await loadModulePromise;
         if (loadModulePromiseResult.isErr()) {
             return err(loadModulePromiseResult.error);
         }
-        const loadModuleResponse = loadModulePromiseResult.value
 
-        return ok(loadModuleResponse);
+        return ok(null);
     }
 
     public async unloadModule(unloadModuleArgs: UnloadModuleArgs): Promise<Result<UnloadModuleResponse,Error>> {
@@ -118,9 +113,9 @@ export class GrpcNodeApiContainerClient implements GenericApiContainerClient {
         return ok(registerServiceResponse)
     }
 
-    public async startService(startServiceArgs: StartServiceArgs): Promise<Result<ServiceInfo, Error>>{
-        const promiseStartService: Promise<Result<ServiceInfo, Error>> = new Promise((resolve, _unusedReject) => {
-            this.client.startService(startServiceArgs, (error: ServiceError | null, response?: ServiceInfo) => {
+    public async startService(startServiceArgs: StartServiceArgs): Promise<Result<StartServiceResponse, Error>>{
+        const promiseStartService: Promise<Result<StartServiceResponse, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.startService(startServiceArgs, (error: ServiceError | null, response?: StartServiceResponse) => {
                 if (error === null) {
                     if (!response) {
                         resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
@@ -132,12 +127,12 @@ export class GrpcNodeApiContainerClient implements GenericApiContainerClient {
                 }
             })
         });
-        const resultStartService: Result<ServiceInfo, Error> = await promiseStartService;
+        const resultStartService: Result<StartServiceResponse, Error> = await promiseStartService;
         if (resultStartService.isErr()) {
             return err(resultStartService.error);
         }
-        const startServiceResponse: ServiceInfo = resultStartService.value;
 
+        const startServiceResponse: StartServiceResponse = resultStartService.value;
         return ok(startServiceResponse)
     }