+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Breaking Changes
+* Revert `StartService` and `LoadModule` changes from 1.52.0
 
 # 1.52.1
 ### Fixes